[Ref] Simplify tone portamento logic. The simplification revealed several small bugs: [Fix] 669: When seeking with sample sync was disabled, the wrong tone portamento amount was memorized to be used by the next note. [Fix] PLM: Tone Portamento memory didn't correctly remember fine slides.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@14949 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/Sndfile.h
@@ -1034,8 +1034,8 @@ protected:
 	void PortamentoExtraFineMPT(ModChannel &chn, int);
 	void SetFinetune(CHANNELINDEX channel, PlayState &playState, bool isSmooth) const;
 	void NoteSlide(ModChannel &chn, uint32 param, bool slideUp, bool retrig) const;
-	std::pair<uint32, bool> GetVolCmdTonePorta(const ModCommand &m, uint32 startTick) const;
-	void TonePortamento(ModChannel &chn, uint32 param) const;
+	std::pair<uint16, bool> GetVolCmdTonePorta(const ModCommand &m, uint32 startTick) const;
+	void TonePortamento(ModChannel &chn, uint16 param) const;
 	void Vibrato(ModChannel &chn, uint32 param) const;
 	void FineVibrato(ModChannel &chn, uint32 param) const;
 	void VolumeSlide(ModChannel &chn, ModCommand::PARAM param);
